Move get_today helper to utils
